docs(status): add comprehensive status update (2026-03-25)
## Summary
- Document cache atomic operations changes
- Report hash output improvement suggestions
- List top 25 tasks to complete
- Document Go cache issue and recovery steps

## Files Added
- docs/status/2026-03-25_20-18_COMPREHENSIVE_STATUS_UPDATE.md

## Files Modified
- cache/file_cache.go: int64 atomic counters
- cache/file_cache_test.go: int64 test assertions

+++ cache/file_cache_test.go
@@ -269,7 +269,7 @@ func TestFileCache_Clear(t *testing.T) {
 }
 
 // assertCacheStats asserts that cache stats match expected values.
-func assertCacheStats(t *testing.T, stats Stats, expectedHits, expectedMisses, expectedSize int) {
+func assertCacheStats(t *testing.T, stats Stats, expectedHits, expectedMisses, expectedSize int64) {
 	t.Helper()
 
 	if stats.Hits != expectedHits {
@@ -280,7 +280,7 @@ func assertCacheStats(t *testing.T, stats Stats, expectedHits, expectedMisses, e
 		t.Errorf("Expected %d misses, got %d", expectedMisses, stats.Misses)
 	}
 
-	if stats.Size != expectedSize {
+	if stats.Size != int(expectedSize) {
 		t.Errorf("Expected %d size, got %d", expectedSize, stats.Size)
 	}
 }
